+++ src/Seis.jl
@@ -110,11 +110,14 @@ export
     # IO
     SACtr,
     channel_code,
+    parse_stationxml,
     read_mseed,
     read_sac,
+    read_stationxml,
     write_mseed,
     write_sac,
     write_sac_header,
+    write_stationxml,
     # Sample data
     sample_data,
     # Rotation
@@ -164,6 +167,8 @@ include("io/SAC/SAC.jl")
 using .SAC
 include("io/Miniseed/Miniseed.jl")
 using .Miniseed
+include("io/StationXML/SeisStationXML.jl")
+using .SeisStationXML
 
 include("show.jl")
 include("input_output.jl")